[ALPS05130148] aee: fix codex issue
1. codex issues

MTK-Commit-Id: dfcab0c188ca7674f8e974042afa594a3faaf7be

CR-Id: ALPS05130148
Feature: Android Exception Engine(AEE)

+++ drivers/misc/mediatek/aee/aed/aed-main.c
@@ -2116,7 +2116,7 @@ int DumpThreadNativeInfo(struct aee_oops *oops)
 	unsigned long userstack_end = 0, length = 0;
 	int mapcount = 0;
 	struct file *file;
-	int flags;
+	unsigned long flags;
 	struct mm_struct *mm;
 	int ret = 0;
 	char tpath[512];
